# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "EDU Api Middleware",
        "description": "A Api Middleware from EDU",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users in System",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create New User On The System",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/users/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the cat that we want to",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User of the given ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    }
                }
            }
        },        
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "Users": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "title": "id"
                },
                "username": {
                    "type": "string",
                    "title": "username"
                },
                "name": {
                    "type": "string",
                    "title": "name"
                },
                "firstName": {
                    "type": "string",
                    "title": "firstName"
                },
                "lastName": {
                    "type": "string",
                    "title": "lastName"
                },
                "email": {
                    "type": "string",
                    "title": "email"
                },
                "password": {
                    "title": "password",
                    "type": "string"
                },
                "confirmPassword": {
                    "title": "password",
                    "type": "boolean"
                },
                "phoneNumber": {
                    "type": "string",
                    "title": "phoneNumber"
                },
                "inactive": {
                    "type": "boolean",
                    "title": "inactive"
                },
                "country": {
                    "type": "string",
                    "title": "country"
                }
            }
        },
...
    }
}